Directory reorganization with added CMake support
