+++ bitwise/include/eteran/cpp-utilities/bitwise.h
