+++ programs/CMakeLists.txt
@@ -1,0 +1,14 @@
+# eteran / cpp-utilities / programs
+cmake_minimum_required(VERSION 3.5)
+
+
+project("eteran-cpp-utilities-programs" CXX)
+add_executable("${PROJECT_NAME}" "")
+
+
+target_sources(
+	"${PROJECT_NAME}"
+	PRIVATE
+		"${CMAKE_CURRENT_LIST_DIR}/byte_writer.cpp"
+)
+
