--- tests/thread_pool.cpp
@@ -1,7 +1,0 @@
-
-#include "../thread_pool.h"
-#include <iostream>
-
-int main() {
-	thread_pool pool;
-}
