+++ thread_pool/test/thread_pool.cpp
@@ -1,0 +1,6 @@
+#include <eteran/cpp-utilities/thread_pool.h>
+#include <iostream>
+
+int main() {
+	thread_pool pool;
+}
